FFWEB-1203 CLI export feed script

+++ src/Export/Output/Csv.php
@@ -48,7 +48,9 @@ class Csv extends AbstractOutput implements OutputInterface
 
     public function __destruct()
     {
-        unlink($this->handle->getPathname());
+        if ($this->handle) {
+            unlink($this->handle->getPathname());
+        }
     }
 
     /**
